requirements.txt
requests
asyncio
kafka-python
pyyaml
datetime
